+++ src/main/java/development/v/development/application/features/CashReceipt/UpdateCashReceipt/UpdateCashReceiptUseCase.java
@@ -54,7 +54,7 @@ public class UpdateCashReceiptUseCase {
         customerRepository.findById(request.getCliId())
                 .orElseThrow(() -> new NotFoundException(Message.NOT_FOUND + ": Cliente " + request.getCliId()));
 
-        orderRepository.findById(request.getPedId())
+        orderRepository.findById(request.getPedId().intValue())
                 .orElseThrow(() -> new NotFoundException(Message.NOT_FOUND + ": Pedido " + request.getPedId()));
     }
 }
